Restrict database operations to superusers and database owner

+++ data-acquisition/app/app/api/api_v1/endpoints/databases.py
@@ -2,6 +2,7 @@ from typing import Any, List
 
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
+from fastapi import Request
 
 from app import crud, models, schemas
 from app.api import deps
@@ -11,6 +12,8 @@ router = APIRouter()
 
 @router.get("/", response_model=List[schemas.Database])
 def read_databases(
+        *,
+        request: Request,
         db: Session = Depends(deps.get_db),
         skip: int = 0,
         limit: int = 100,
@@ -18,13 +21,24 @@ def read_databases(
     """
     Retrieve databases.
     """
-    databases = crud.database.get_multi(db, skip=skip, limit=limit)
+    query_params = dict(request.query_params.items())
+    for key in ['skip', 'limit']:
+        query_params.pop(key)
+    user = models.User(**query_params)
+
+    if user.is_superuser:
+        databases = crud.database.get_multi(db, skip=skip, limit=limit)
+    else:
+        databases = crud.database.get_multi_by_owner(
+            db=db, owner_id=user.id, skip=skip, limit=limit
+        )
     return databases
 
 
 @router.post("/", response_model=schemas.Database)
 def create_database(
         *,
+        request: Request,
         db: Session = Depends(deps.get_db),
         database_in: schemas.DatabaseCreate,
 ) -> Any:
@@ -38,6 +52,7 @@ def create_database(
 @router.put("/{id}", response_model=schemas.Database)
 def update_database(
         *,
+        request: Request,
         db: Session = Depends(deps.get_db),
         id: int,
         database_in: schemas.DatabaseUpdate,
@@ -45,9 +60,16 @@ def update_database(
     """
     Update a database.
     """
+    query_params = dict(request.query_params.items())
+    for key in ['id']:
+        query_params.pop(key)
+    user = models.User(**query_params)
+
     database = crud.database.get(db=db, id=id)
     if not database:
         raise HTTPException(status_code=404, detail="Database not found")
+    if not user.is_superuser and (database.owner_id != user.id):
+        raise HTTPException(status_code=400, detail="Not enough permissions")
     database = crud.database.update(db=db, db_obj=database, obj_in=database_in)
     return database
 
@@ -55,29 +77,45 @@ def update_database(
 @router.get("/{id}", response_model=schemas.Database)
 def read_database(
         *,
+        request: Request,
         db: Session = Depends(deps.get_db),
         id: int,
 ) -> Any:
     """
     Get database by ID.
     """
+    query_params = dict(request.query_params.items())
+    for key in ['id']:
+        query_params.pop(key)
+    user = models.User(**query_params)
+
     database = crud.database.get(db=db, id=id)
     if not database:
         raise HTTPException(status_code=404, detail="Database not found")
+    if not user.is_superuser and (database.owner_id != user.id):
+        raise HTTPException(status_code=400, detail="Not enough permissions")
     return database
 
 
 @router.delete("/{id}", response_model=schemas.Database)
 def delete_database(
         *,
+        request: Request,
         db: Session = Depends(deps.get_db),
         id: int,
 ) -> Any:
     """
     Delete a database.
     """
+    query_params = dict(request.query_params.items())
+    for key in ['id']:
+        query_params.pop(key)
+    user = models.User(**query_params)
+
     database = crud.database.get(db=db, id=id)
     if not database:
         raise HTTPException(status_code=404, detail="Database not found")
+    if not user.is_superuser and (database.owner_id != user.id):
+        raise HTTPException(status_code=400, detail="Not enough permissions")
     database = crud.database.remove(db=db, id=id)
     return database
